+++ app/src/main/java/com/kuit/ourmenu/ui/navigator/Routes.kt
@@ -26,11 +26,11 @@ sealed interface Routes{
 
     // Onboarding
     @Serializable
-    data object Signup: Routes
-    @Serializable
     data object Landing: Routes
     @Serializable
     data object Login: Routes
+
+    // Signup
     @Serializable
     data object SignupEmail: Routes
     @Serializable
