+++ app/src/main/java/com/kuit/ourmenu/ui/signup/screen/SignupMealTimeScreen.kt
@@ -1,4 +1,4 @@
-package com.kuit.ourmenu.ui.onboarding.screen.signup
+package com.kuit.ourmenu.ui.signup.screen
 
 import android.util.Log
 import androidx.compose.foundation.layout.Box
@@ -22,7 +22,7 @@ import com.kuit.ourmenu.R
 import com.kuit.ourmenu.ui.common.DisableBottomFullWidthButton
 import com.kuit.ourmenu.ui.onboarding.component.MealTimeGrid
 import com.kuit.ourmenu.ui.onboarding.component.OnboardingTopAppBar
-import com.kuit.ourmenu.ui.onboarding.state.SignupState
+import com.kuit.ourmenu.ui.signup.model.SignupState
 import com.kuit.ourmenu.ui.onboarding.viewmodel.SignupViewModel
 import com.kuit.ourmenu.ui.theme.Neutral500
 import com.kuit.ourmenu.ui.theme.Neutral900
